+++ CHANGES.markdown
@@ -1,3 +1,6 @@
+## Changes in 2.6.1
+   - Add `Test.Hspec.Runner.runSpec` (see #216)
+
 ## Changes in 2.6.0
    - Allow to focus individual spec items (see #319)
    - Do not calculate diff on `--no-diff` (see #332)